+++ cpp/tests/datetime/datetime_ops_test.cpp
@@ -914,76 +914,4 @@ TYPED_TEST(TypedDatetimeOpsTest, TestFloorDatetime)
   CUDF_TEST_EXPECT_COLUMNS_EQUAL(*floor_millisecond(input), expected_nanosecond);
 }
 
-TYPED_TEST(TypedDatetimeOpsTest, TestRoundDatetime)
-{
-  using T = TypeParam;
-  using namespace cudf::test;
-  using namespace cudf::datetime;
-  using namespace cuda::std::chrono;
-
-  auto start = milliseconds(-2500000000000);  // Sat, 11 Oct 1890 19:33:20 GMT
-  auto stop  = milliseconds(2500000000000);   // Mon, 22 Mar 2049 04:26:40 GMT
-
-  auto input = generate_timestamps<T>(this->size(), time_point_ms(start), time_point_ms(stop));
-
-  auto host_val                     = to_host<T>(input);
-  thrust::host_vector<T> timestamps = host_val.first;
-
-  std::vector<T> round_day(timestamps.size());
-  std::transform(timestamps.begin(), timestamps.end(), round_day.begin(), [](auto i) {
-    return time_point_cast<typename T::duration>(round<days>(i));
-  });
-  auto expected_day =
-    fixed_width_column_wrapper<T, typename T::duration::rep>(round_day.begin(), round_day.end());
-  CUDF_TEST_EXPECT_COLUMNS_EQUAL(*round_day(input), expected_day);
-
-  std::vector<T> round_hour(timestamps.size());
-  std::transform(timestamps.begin(), timestamps.end(), round_hour.begin(), [](auto i) {
-    return time_point_cast<typename T::duration>(round<hours>(i));
-  });
-  auto expected_hour =
-    fixed_width_column_wrapper<T, typename T::duration::rep>(round_hour.begin(), round_hour.end());
-  CUDF_TEST_EXPECT_COLUMNS_EQUAL(*round_hour(input), expected_hour);
-
-  std::vector<T> round_minute(timestamps.size());
-  std::transform(timestamps.begin(), timestamps.end(), round_minute.begin(), [](auto i) {
-    return time_point_cast<typename T::duration>(round<minutes>(i));
-  });
-  auto expected_minute = fixed_width_column_wrapper<T, typename T::duration::rep>(
-    round_minute.begin(), round_minute.end());
-  CUDF_TEST_EXPECT_COLUMNS_EQUAL(*round_minute(input), expected_minute);
-
-  std::vector<T> round_second(timestamps.size());
-  std::transform(timestamps.begin(), timestamps.end(), round_second.begin(), [](auto i) {
-    return time_point_cast<typename T::duration>(round<seconds>(i));
-  });
-  auto expected_second = fixed_width_column_wrapper<T, typename T::duration::rep>(
-    round_second.begin(), round_second.end());
-  CUDF_TEST_EXPECT_COLUMNS_EQUAL(*round_second(input), expected_second);
-
-  std::vector<T> round_millisecond(timestamps.size());
-  std::transform(timestamps.begin(), timestamps.end(), round_millisecond.begin(), [](auto i) {
-    return time_point_cast<typename T::duration>(round<milliseconds>(i));
-  });
-  auto expected_millisecond = fixed_width_column_wrapper<T, typename T::duration::rep>(
-    round_millisecond.begin(), round_millisecond.end());
-  CUDF_TEST_EXPECT_COLUMNS_EQUAL(*round_millisecond(input), expected_millisecond);
-
-  std::vector<T> round_microsecond(timestamps.size());
-  std::transform(timestamps.begin(), timestamps.end(), round_microsecond.begin(), [](auto i) {
-    return time_point_cast<typename T::duration>(round<microseconds>(i));
-  });
-  auto expected_microsecond = fixed_width_column_wrapper<T, typename T::duration::rep>(
-    round_microsecond.begin(), round_microsecond.end());
-  CUDF_TEST_EXPECT_COLUMNS_EQUAL(*round_second(input), expected_microsecond);
-
-  std::vector<T> round_nanosecond(timestamps.size());
-  std::transform(timestamps.begin(), timestamps.end(), round_nanosecond.begin(), [](auto i) {
-    return time_point_cast<typename T::duration>(round<nanoseconds>(i));
-  });
-  auto expected_nanosecond = fixed_width_column_wrapper<T, typename T::duration::rep>(
-    round_nanosecond.begin(), rounded_nanosecond.end());
-  CUDF_TEST_EXPECT_COLUMNS_EQUAL(*round_millisecond(input), expected_nanosecond);
-}
-
 CUDF_TEST_PROGRAM_MAIN()
