+++ docs/cudf/source/api_docs/index_objects.rst
@@ -282,7 +282,6 @@ Time-specific operations
    DatetimeIndex.round
    DatetimeIndex.ceil
    DatetimeIndex.floor
-   DatetimeIndex.round
 
 Conversion
 ~~~~~~~~~~
