+++ docs/cudf/source/api_docs/series.rst
@@ -303,7 +303,6 @@ Datetime methods
    isocalendar
    ceil
    floor
-   round
 
 
 Timedelta properties
