+++ python/cudf/cudf/_lib/cpp/datetime.pxd
@@ -39,19 +39,6 @@ cdef extern from "cudf/datetime.hpp" namespace "cudf::datetime" nogil:
     cdef unique_ptr[column] floor_nanosecond(
         const column_view& column
     ) except +
-    cdef unique_ptr[column] round_day(const column_view& column) except +
-    cdef unique_ptr[column] round_hour(const column_view& column) except +
-    cdef unique_ptr[column] round_minute(const column_view& column) except +
-    cdef unique_ptr[column] round_second(const column_view& column) except +
-    cdef unique_ptr[column] round_millisecond(
-        const column_view& column
-    ) except +
-    cdef unique_ptr[column] round_microsecond(
-        const column_view& column
-    ) except +
-    cdef unique_ptr[column] round_nanosecond(
-        const column_view& column
-    ) except +
     cdef unique_ptr[column] add_calendrical_months(
         const column_view& timestamps,
         const column_view& months
