+++ python/cudf/cudf/_lib/datetime.pyx
@@ -116,33 +116,6 @@ def floor_datetime(Column col, object field):
     return result
 
 
-def round_datetime(Column col, object field):
-    cdef unique_ptr[column] c_result
-    cdef column_view col_view = col.view()
-
-    with nogil:
-        # https://pandas.pydata.org/docs/reference/api/pandas.Timedelta.resolution_string.html
-        if field == "D":
-            c_result = move(libcudf_datetime.round_day(col_view))
-        elif field == "H":
-            c_result = move(libcudf_datetime.round_hour(col_view))
-        elif field == "T" or field == "min":
-            c_result = move(libcudf_datetime.round_minute(col_view))
-        elif field == "S":
-            c_result = move(libcudf_datetime.round_second(col_view))
-        elif field == "L" or field == "ms":
-            c_result = move(libcudf_datetime.round_millisecond(col_view))
-        elif field == "U" or field == "us":
-            c_result = move(libcudf_datetime.round_microsecond(col_view))
-        elif field == "N":
-            c_result = move(libcudf_datetime.round_nanosecond(col_view))
-        else:
-            raise ValueError(f"Invalid resolution: '{field}'")
-
-    result = Column.from_unique_ptr(move(c_result))
-    return result
-
-
 def is_leap_year(Column col):
     """Returns a boolean indicator whether the year of the date is a leap year
     """
