+++ python/cudf/cudf/tests/test_datetime.py
@@ -1819,34 +1819,3 @@ def test_floor(data, time_type, resolution):
     expect = ps.dt.floor(resolution)
     got = gs.dt.floor(resolution)
     assert_eq(expect, got)
-
-
-@pytest.mark.parametrize(
-    "data",
-    [
-        (
-            [
-                "2020-05-31 08:00:00",
-                "1999-12-31 18:40:10",
-                "2000-12-31 04:00:05",
-                "1900-02-28 07:00:06",
-                "1800-03-14 07:30:20",
-                "2100-03-14 07:30:20",
-                "1970-01-01 00:00:09",
-                "1969-12-31 12:59:10",
-            ]
-        )
-    ],
-)
-@pytest.mark.parametrize("time_type", DATETIME_TYPES)
-@pytest.mark.parametrize(
-    "resolution", ["D", "H", "T", "min", "S", "L", "ms", "U", "us", "N"]
-)
-def test_round(data, time_type, resolution):
-
-    gs = cudf.Series(data, dtype=time_type)
-    ps = gs.to_pandas()
-
-    expect = ps.dt.round(resolution)
-    got = gs.dt.round(resolution)
-    assert_eq(expect, got)
