+++ python/cudf/cudf/tests/test_index.py
@@ -2496,16 +2496,3 @@ def test_index_datetime_floor(resolution):
     cuidx_floor = cuidx.floor(resolution)
 
     assert_eq(pidx_floor, cuidx_floor)
-
-
-@pytest.mark.parametrize(
-    "resolution", ["D", "H", "T", "min", "S", "L", "ms", "U", "us", "N"]
-)
-def test_index_datetime_round(resolution):
-    cuidx = cudf.DatetimeIndex([1000000, 2000000, 3000000, 4000000, 5000000])
-    pidx = cuidx.to_pandas()
-
-    pidx_floor = pidx.round(resolution)
-    cuidx_floor = cuidx.round(resolution)
-
-    assert_eq(pidx_floor, cuidx_floor)
